ottimizzazione utilizzo semafori

+++ src/alimentation.c
@@ -39,7 +39,7 @@ int main(int argc, char *argv[]){
 	signal_handler_init();
 	shm_info_attach(&stats.info);
 	sem_execute_semop(shm_sem_get_startid(stats.info), 0, 1, 0);
-	while(sem_getval(shm_sem_get_startid(stats.info), 1) != (long)1){
+	while(sem_getval(shm_sem_get_startid(stats.info), 1) != 1){
 		sleep(1);
 	}
 	step_nsec = shm_info_get_step(stats.info);	//ogni quanti nano secondi creare n atomi
